refactor(dispatch): extract envelope_govern.py leaf module (PR-4 monolith split)
Moves _govern (346 lines, unchanged) from dispatch_envelope.py to
scripts/lib/envelope_govern.py. dispatch_envelope.py re-imports it at bare
name so every existing dispatch_envelope._govern-shaped coupling keeps
binding — its 8 call-site patches (test_dispatch_envelope_field_propagation.py,
test_dispatch_envelope_plan.py) stay green untouched, since run_envelope /
run_envelope_plan / run_envelope_headless_plan all stay in the facade.

_govern is the CALLER of _archive_dispatch_events and _clear_dispatch_events
(moved to envelope_govern_support.py in PR-3). Moving the caller re-targets
every patch("dispatch_envelope._archive_dispatch_events"/"_clear_dispatch_events")
coupling onto envelope_govern.<name> — six sites across three files:

  - tests/test_dispatch_envelope.py:236-237 (asserted at 248)
  - tests/test_dispatch_envelope_characterization.py:234-235 (asserted)
  - tests/test_dispatch_envelope_characterization.py:288-289 (peripheral stub)
  - tests/test_role_applied_provider_lane.py:211-212 (previously unasserted —
    added explicit mock_archive/mock_clear assertions so an unbound patch now
    fails loudly instead of silently running the real event-store teardown)

Census fixture regenerated (tests/dispatch_envelope_census_scanner.py) to
reflect the new envelope_govern family member and the six re-targeted
couplings.

Dispatch-ID: 20260803-225854-pr4-envelope-govern
